requirements.txt
langgraph
langchain
langchain-openai
pydantic
fastapi
uvicorn
httpx
python-dotenv
pytest
pymupdf
pypdf
langchain-community
qdrant-client
mistralai
pandas
tabulate
lxml